refactor: Update imports to use FCEApplicationForm types and remove unused education-details-dialog component
- Refactor applications-table and education-details-dialog components to import types from FCEApplicationForm instead of ApplicationForm, aligning with the new structure.
- Delete the unused education-details-dialog component from the CRM payload, improving code maintainability and reducing clutter.
